double to int

+++ DesafioMusica/br/com/alura/desafiomusica/modelos/Audio.java
@@ -5,7 +5,7 @@ public class Audio {
     private String titulo;
     private int totalReproducoes;
     private int totalCurtidas;
-    private double classificacao;
+    private int classificacao;
 
     public String getTitulo() {
         return titulo;
@@ -31,11 +31,11 @@ public class Audio {
         this.totalCurtidas = totalCurtidas;
     }
 
-    public double getClassificacao() {
+    public int getClassificacao() {
         return classificacao;
     }
 
-    public void setClassificacao(double classificacao) {
+    public void setClassificacao(int classificacao) {
         this.classificacao = classificacao;
     }
